refactor(ui): merge two SummaryCard duplicates into shared/SummaryCard
pages/bill-calendar/components/SummaryCard.tsx and pages/subscription-tracker/components/SummaryCard.tsx were 95% identical -- same icon + label + value layout, same prop names, same animation. Differences: bill-calendar used flat p-6 padding, subscription-tracker used responsive p-4 md:p-6, and bill-calendar truncated long values.

New shared SummaryCard at components/shared/SummaryCard.tsx absorbs both. Compact prop preserves the responsive p-4 md:p-6 padding (subscription-tracker passes compact, bill-calendar uses default p-6). truncate + min-w-0 always-on so long values clip cleanly across both callsites.

Two old files deleted; both pages updated to import from shared. Visual output unchanged where each page already worked. tsc + eslint + build clean.
